requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
logging==0.4.9.6
multidict==6.1.0
mutagen==1.47.0
mypy==1.13.0
mypy-extensions==1.0.0
propcache==0.2.0
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.16.0
youtube-dl==2021.12.17
yt-dlp==2024.10.22
